+++ src/__tests__/middlewares/validationMiddlewares/ordersValidations.test.ts
@@ -1,5 +1,5 @@
 
-import { validateCheckOrder,  validateOrderId, validateOrdersStatus, } from "../../../middlewares/validationsFunctions";
+import { validateCheckOrder,  validateOrderId, validateOrdersStatus, } from "../../../middlewares/validationFunctions/ordersValidationFunctions";
 import { hexWith24Char, hexWith25Char, stringWith24Char } from "../../assets/testData/stringTestData";
 import { createResponseNext, expectValidationError, expectValidationPassed } from "../../utils/helperTestFunctions.test";
 import {Request} from "express";
