+++ src/__tests__/middlewares/validationMiddlewares/productsValidations.test.ts
@@ -1,11 +1,13 @@
 import { emptyImageObj, wrongDataTypesImage, correctImageObj } from './../../assets/testData/fileTestData';
 import Joi from "joi";
-import { validateCreateProduct, validateUpdateProduct, validateUpdateSingleImageProduct, } from "../../../middlewares/validationsFunctions";
+import { validateCreateProduct, validateUpdateProduct, validateUpdateSingleImageProduct,
+ } from "../../../middlewares/validationFunctions/productsValidationFunctions";
 import {  hexWith23Char, hexWith24Char, hexWith25Char, maxOffer, maxPrice, minOffer, minPrice,
      stringWith100Char, stringWith101Char, stringWith1024Char, stringWith1025Char, stringWith10Char, 
-     stringWith24Char,  stringWith2Char, stringWith32Char, stringWith3Char, stringWith5Char, stringWith9Char }
+     stringWith24Char,  stringWith2Char, stringWith32Char, stringWith33Char, stringWith3Char, stringWith5Char, stringWith9Char }
       from "../../assets/testData/stringTestData";
-import { createResponseNext, expectValidationError, expectValidationPassed, extractJoiCallErrorMessage } from "../../utils/helperTestFunctions.test";
+import { createResponseNext, expectValidationError, expectValidationPassed,
+     extractJoiCallErrorMessage } from "../../utils/helperTestFunctions.test";
 import {Request} from "express";
 
 
@@ -32,7 +34,7 @@ describe("Products validation middlewares",()=>{
             expect(next).toHaveBeenCalledTimes(1);
         })
 
-        it("Should return an error with all fields below minimum length or number or when brand is not set to the valid options",()=>{
+        it("Should return an error with all fields below minimum length or number",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -58,12 +60,11 @@ describe("Products validation middlewares",()=>{
                 "price must be greater than or equal to 0",
                 "finalPrice must be greater than or equal to 0",
                 "quantity must be greater than or equal to 0",
-                "brand must be one of [Nike, Levis, Calvin Klein, Casio, Adidas, Biba]",
             ])
             expect(next).toHaveBeenCalledTimes(1);
         })
 
-        it("Should return an error with all fields above length or number or when brand is not set to the valid options",()=>{
+        it("Should return an error with all fields above length or number allowed",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -86,12 +87,11 @@ describe("Products validation middlewares",()=>{
                 "offer must be less than or equal to 1",
                 "price must be less than or equal to 1000",
                 "finalPrice must be less than or equal to 1000",
-                "brand must be one of [Nike, Levis, Calvin Klein, Casio, Adidas, Biba]",
             ])
             expect(next).toHaveBeenCalledTimes(1);
         })
 
-        it("Should pass when all the parameters are set to the maximum allowed and brand is set to allowed strings",()=>{
+        it("Should pass when all the parameters are set to the maximum allowed",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -101,7 +101,7 @@ describe("Products validation middlewares",()=>{
                     offer:maxOffer,
                     price:maxPrice,
                     finalPrice:maxPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateCreateProduct(req,res,next);
@@ -119,7 +119,7 @@ describe("Products validation middlewares",()=>{
                     offer:minOffer,
                     price:minPrice,
                     finalPrice:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateCreateProduct(req,res,next);
@@ -136,7 +136,7 @@ describe("Products validation middlewares",()=>{
                     categoryId:hexWith24Char,
                     offer:0.5,
                     finalPrice:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateCreateProduct(req,res,next);
@@ -159,7 +159,7 @@ describe("Products validation middlewares",()=>{
                     offer:0.5,
                     quantity:1.2,
                     price:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateCreateProduct(req,res,next);
@@ -182,7 +182,7 @@ describe("Products validation middlewares",()=>{
                     offer:0.5,
                     quantity:1,
                     price:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateCreateProduct(req,res,next);
@@ -223,7 +223,7 @@ describe("Products validation middlewares",()=>{
         //     expect(next).toHaveBeenCalledTimes(1);
         // })
 
-        it("Should return an error with all fields below minimum length or number or when brand is not set to the valid options",()=>{
+        it("Should return an error with all fields below minimum length or number or when brand is empty",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -234,7 +234,7 @@ describe("Products validation middlewares",()=>{
                     price:minPrice - 1,
                     finalPrice:-1,
                     quantity:-1,
-                    brand:stringWith5Char
+                    brand:""
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
@@ -249,12 +249,12 @@ describe("Products validation middlewares",()=>{
                 "price must be greater than or equal to 0",
                 "finalPrice must be greater than or equal to 0",
                 "quantity must be greater than or equal to 0",
-                "brand must be one of [Nike, Levis, Calvin Klein, Casio, Adidas, Biba]",
+                "brand is not allowed to be empty",
             ])
             expect(next).toHaveBeenCalledTimes(1);
         })
 
-        it("Should return an error with all fields above length or number or when brand is not set to the valid options",()=>{
+        it("Should return an error with all fields above length or number allowed",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -264,7 +264,7 @@ describe("Products validation middlewares",()=>{
                     offer:maxOffer + 0.01,
                     price:maxPrice + 1,
                     finalPrice:maxPrice + 1,
-                    brand:stringWith5Char
+                    brand:stringWith33Char
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
@@ -277,12 +277,12 @@ describe("Products validation middlewares",()=>{
                 "offer must be less than or equal to 1",
                 "price must be less than or equal to 1000",
                 "finalPrice must be less than or equal to 1000",
-                "brand must be one of [Nike, Levis, Calvin Klein, Casio, Adidas, Biba]",
+                "brand length must be less than or equal to 32 characters long",
             ])
             expect(next).toHaveBeenCalledTimes(1);
         })
 
-        it("Should pass when all the parameters are set to the maximum allowed and brand is set to allowed strings",()=>{
+        it("Should pass when all the parameters are set to the maximum allowed",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -292,7 +292,7 @@ describe("Products validation middlewares",()=>{
                     offer:maxOffer,
                     price:maxPrice,
                     finalPrice:maxPrice,
-                    brand:"Nike"
+                    brand:stringWith2Char
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
@@ -300,7 +300,7 @@ describe("Products validation middlewares",()=>{
             expectValidationPassed(next)
         })
 
-        it("Should pass when all the parameters are set to the minimum allowed and brand is set to allowed strings",()=>{
+        it("Should pass when all the parameters are set to the minimum allowed",()=>{
            const { next,res } = createResponseNext()
             const req = {
                 body:{
@@ -310,7 +310,7 @@ describe("Products validation middlewares",()=>{
                     offer:minOffer,
                     price:minPrice,
                     finalPrice:minPrice,
-                    brand:"Nike"
+                    brand:stringWith2Char
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
@@ -331,7 +331,7 @@ describe("Products validation middlewares",()=>{
                     categoryId:hexWith24Char,
                     offer:0.5,
                     finalPrice:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
@@ -349,7 +349,7 @@ describe("Products validation middlewares",()=>{
                     offer:0.5,
                     quantity:1.2,
                     price:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
@@ -372,7 +372,7 @@ describe("Products validation middlewares",()=>{
                     offer:0.5,
                     quantity:1,
                     price:minPrice,
-                    brand:"Nike"
+                    brand:stringWith5Char
                 }
             } as Request;
             validateUpdateProduct(req,res,next);
