+++ src/middlewares/pagination.ts
@@ -1,19 +1,25 @@
 import { NextFunction, Request, Response } from "express";
 
-export const pagination = (req:Request,res:Response,next:NextFunction)=>{
-    let limit = Number(parseInt(req.query.limit as string));
-    let page = Number(parseInt(req.query.page as string));
 
-    if(Number.isNaN(page) || Number.isNaN(limit) || limit > 30 || limit <= 0 || page <= 0){
-        limit = 9;
-        page = 1;
-    }
-    let skip = limit  * (page - 1);
+const pagination = (minLimit:number=0,maxLimit:number=30)=>{
+    return (req:Request,res:Response,next:NextFunction)=>{
+        let limit = Number(parseInt(req.query.limit as string));
+        let page = Number(parseInt(req.query.page as string));
 
-    req.pagination = {
-        limit,
-        page,
-        skip,
+        if(Number.isNaN(page) || Number.isNaN(limit) || limit > maxLimit || limit <= minLimit || page <= 0){
+            limit = 9;
+            page = 1;
+        }
+        let skip = limit  * (page - 1);
+
+        req.pagination = {
+            limit,
+            page,
+            skip,
+        }
+        next();
     }
-    next();
 }
+
+export const paginationUser = pagination(1,30);
+export const paginationAdmin = pagination(1,100);
