+++ src/middlewares/validationSchemas/sharedValidationSchemas.ts
@@ -1,0 +1,10 @@
+import Joi from "joi";
+
+export const fileImageSchema = Joi.object({
+    fieldname: Joi.string().required(),
+    originalname: Joi.string().required(),
+    encoding: Joi.string().required(),
+    mimetype: Joi.string().required(),
+    size: Joi.number().required(),
+    buffer: Joi.binary().required()
+})
