+++ src/routes/addressesRoute.ts
@@ -1,7 +1,7 @@
 import { authenticateUser } from './../middlewares/authenticate';
 import * as addressesRouter from "../controllers/addressesController"
 import express from "express";
-import { validateCreatingAddress, validateUpdatingAddress } from "../middlewares/validationsFunctions";
+import { validateCreatingAddress, validateUpdatingAddress } from "../middlewares/validationFunctions/addressesValidationFunctions";
 const router = express.Router()
 
 router.post("/",validateCreatingAddress,authenticateUser,addressesRouter.createNewAddress)
