+++ src/routes/productsRoute.ts
@@ -1,17 +1,17 @@
-import { pagination } from './../middlewares/pagination';
+import { paginationUser } from './../middlewares/pagination';
 import express from "express"
 import * as productsRouter from "../controllers/productsController"
 import { authenticateAdmin } from '../middlewares/authenticate';
-import {  validateAppendImagesToProduct, validateCreateProduct, validateUpdateProduct, validateUpdateSingleImageProduct } from '../middlewares/validationsFunctions';
 import upload from '../utils/Multer';
+import { validateAppendImagesToProduct, validateCreateProduct, validateUpdateProduct, validateUpdateSingleImageProduct } from '../middlewares/validationFunctions/productsValidationFunctions';
 const MB2 = 2000;
 const MB10 = 10000;
 const MB12 = 12000;
 // import { getCache } from '../middlewares/cache';
 const router = express.Router()
 
-router.get("/:productId",pagination,productsRouter.getProductById);
-router.get("/",pagination,productsRouter.getAllProducts);
+router.get("/:productId",paginationUser,productsRouter.getProductById);
+router.get("/",paginationUser,productsRouter.getAllProducts);
 router.get("/search/:text",productsRouter.searchForProducts);
 router.post("/",authenticateAdmin, upload({fileSize:MB2}).single('image'),validateCreateProduct,productsRouter.postNewProduct);
 router.patch("/:productId",authenticateAdmin,upload({}).none(),validateUpdateProduct, productsRouter.updateProductInfo);
