+++ src/routes/reviewsRoute.ts
@@ -1,12 +1,13 @@
-import { authenticateUser, authenticateAdmin } from './../middlewares/authenticate';
-import { pagination } from './../middlewares/pagination';
+import { authenticateUser } from './../middlewares/authenticate';
+import { paginationUser } from './../middlewares/pagination';
 import * as reviewsRouter from "../controllers/reviewsController"
 import express from "express"
-import { validateDeleteUserReview, validateEditUserReview, validateUserReview } from '../middlewares/validationsFunctions';
+import { validateDeleteUserReview, validateEditUserReview, 
+    validateUserReview } from '../middlewares/validationFunctions/reviewsValidationFunctions';
 const router = express.Router()
 
 
-router.get("/:userId",authenticateUser,pagination,reviewsRouter.getAllReviewsByUserId);
+router.get("/:userId",authenticateUser,paginationUser,reviewsRouter.getAllReviewsByUserId);
 router.post("/",validateUserReview,authenticateUser,reviewsRouter.addReviewToProduct);
 router.put("/:reviewId",validateEditUserReview,authenticateUser,reviewsRouter.editReview);
 router.delete("/:productId/:reviewId",validateDeleteUserReview,authenticateUser,reviewsRouter.deleteReview);
